[SPARK-44341][SQL][PYTHON][FOLLOWUP] Move the base trait WindowEvaluatorFactoryBase to a single file
### What changes were proposed in this pull request?
https://github.com/apache/spark/pull/41939 defined the computing logic through PartitionEvaluator API and use it in `WindowExec` and `WindowInPandasExec`.

According to the comment https://github.com/apache/spark/pull/41939#discussion_r1270194752, this PR want move the base trait `WindowEvaluatorFactoryBase` to a single file.

### Why are the changes needed?
Improve the code.

### Does this PR introduce _any_ user-facing change?
'No'.
Just update inner implementation.

### How was this patch tested?
N/A

Closes #42106 from beliefer/SPARK-44341_followup.

Authored-by: Jiaan Geng <beliefer@163.com>
